Add files via upload
The subject notebook is to predict the different delay types should the flight get delayed. Due to the low percentage of the security delay and weather delay, the imblearn library has been used to rebalance the data, however it has found that the result is not significant. Different classification models have been tried in this classification problem, and XGBoost has achieved the best result.